Finsihed implementing the Database Setup functionality and approprioate views. Created SelectThenActOnItem and the SetThenActOnItem superclasses to be used with ComboBoxDialogView and TextBoxDialogView respectively. This is just in case I need to reuse some selection dialogs. Also deriving ClearDatabase and DeleteDatabase ViewModels from ChooseDatabaseVIewModel did not seem clean. The following issues are still open with Sqlite:\ 1. Clearing database doesn't change filesize 2. Deleting database cause file sharing exception when database was in use during the same session These have been documented in the notes and will have to be dealt with later.
I decided to let the connection string of SQlite contain the entire Path for the databse.
All workarounds to just save database name seemed awkward. \
It is easier to just create the "pokertell" database at the first run of PokerTell or let the user pick a name for it.

STarted to implemnt DatabaseConnector.ConnectToDatabase. TryToConnectToDatabaseUsing() is basically finsihed.
In the process I created the StatusUpdateEvent class and appropriate Event args. In this case it will be used to trigger the status bar update reflecting the newly established database connection.

915 tests pass.


git-svn-id: http://svn.xp-dev.com/svn/Software@44 8eca4967-6344-411a-a426-5af860f6c464