Implemented ConnectToDatabase on Database Connector. Added RepositoryDatabase, ConvertedPokerHandInserter - and Retriever to the Repository. Added Repository Insert and Retrieve hand(s) methods which forward work to above mentioned classes. These methods were tested but the classes itself not, since they were just reused and cleaned up from the Legacy code in which they had worked without any problems. Test may be added later though, therefore I made sure to inject all dependencies including DatabaseUtility which I chose to not make static so it can implement an interface and be injected were needed.
Then I created a StatusBarView in the User which right now picks up the DatabaseStatusUpdate Event raised by the DatabaseConnector.
When choosing a database this needs to be raised as well.
Another event like dtaabase connection (DataProvider changed) established needs to be raised as well, so the Repository can have its DataProvider updated.

Since menus kept appearing behind the Window, I implemented a simple GetOsVersion function. If the OS is XP or older, I automatically forbid Transparency for all Windows that contain an expanding ItemsControl (e.g. Menu or ComboBox).
I think this simple solution also make it less of a hassle for the user and it doesn't look so bad after all.

927 tests pass


git-svn-id: http://svn.xp-dev.com/svn/Software@45 8eca4967-6344-411a-a426-5af860f6c464