Adjusted model to compply more witht he one suggested in NHibernate in Action. Therfore I created a SessionFactoryManager which is used by the Daos and TransactionManager to pull the current sesssion. It is responsible for keeping track of the current DataProvider and to hold a SessionFactory buildt from it. The latter also has now the added responsiblity to bind/unbind the current session. It now also has a BatchExecute which uses a stateless session and an Execute thta takes a function and returns its output when executed. Only during the normal Execute is the session bound/unbound and a transaction started and committed. During BatchExecute a transaction on the stateless session is started and committed.
SQLite was fully integrated some minor changes were necessary to tweak the sql-types used (it is not so smart with SQLite).
MOst importantly the GameId column had to be set to bigint(20) otherwise an integer is used and since they are all to big, the values get stored as 0s instead.

The necessary changes to the RepositoryModule have been made.

The application has been run and tested with above changes and efficiency measured.
It is still at about 22000 hands / 17 min (MySql) - SQLite has not been exactly measured but is similar.

1017 tests pass.


git-svn-id: http://svn.xp-dev.com/svn/Software@53 8eca4967-6344-411a-a426-5af860f6c464